talkBotv3/requirements.txt
async_timeout==4.0.2
discord.py==2.3.2
elevenlabs==1.2.2
python-dotenv==1.0.1
